+++ dict_to_list.py
@@ -1,0 +1,10 @@
+# to convert nested dictionary JSON response into a list of values
+def get_all_values(d):
+    if isinstance(d, dict):
+        for v in d.values():
+            yield from get_all_values(v)
+    elif isinstance(d, list):
+        for v in d:
+            yield from get_all_values(v)
+    else:
+        yield d
