+++ main.py
@@ -2,6 +2,8 @@ from googleapiclient.discovery import build
 
 import os
 
+from dict_to_list import get_all_values
+
 # paste the API key generated by you here
 
 # The API Key here is hidden by saving to an environment variable.
@@ -10,19 +12,6 @@ API_KEY = os.environ.get('API_KEY')
 # Sana, Fauna, Kronii, Mumei, Baelz Channel URLs
 channel_url_id = ["UCsUj0dszADCGbF3gNrQEuSQ", "UCO_aKKYxn4tvrqPjcTzZ6EQ",
               "UCmbs8T6MWqUHP1tIQvSgKrg", "UC3n5uGu18FoCy23ggWWp8tA", "UCgmPnx-EEeOrZSg5Tiw7ZRQ"]
-
-
-# to convert nested dictionary JSON response into a list of values
-def get_all_values(d):
-    if isinstance(d, dict):
-        for v in d.values():
-            yield from get_all_values(v)
-    elif isinstance(d, list):
-        for v in d:
-            yield from get_all_values(v)
-    else:
-        yield d
-
 
 # for building a request to the YouTube API using an API Key.
 youtube = build('youtube', 'v3', developerKey=API_KEY)
